Use size_t instead of int to support arrays of more than 2G elements

+++ deps/cuda01.jl
@@ -8,8 +8,8 @@ function cuda01src(f, j=f, ex="$f(xi,yi)"; BLK=256, THR=256)
     for (T,F) in [("float","$(f)_32"),("double","$(f)_64")]
         print(s,
 """
-__global__ void _$(F)_01(int n, $T xi, $T *y, $T *z) {
-  int i = threadIdx.x + blockIdx.x * blockDim.x;
+__global__ void _$(F)_01(size_t n, $T xi, $T *y, $T *z) {
+  size_t i = threadIdx.x + blockIdx.x * blockDim.x;
   while (i < n) {
     $T yi = y[i];
     z[i] = $ex;
@@ -17,7 +17,7 @@ __global__ void _$(F)_01(int n, $T xi, $T *y, $T *z) {
   }
 }
 extern "C" {
-  $DLLEXPORT void $(F)_01(int n, $T xi, $T *y, $T *z) {
+  $DLLEXPORT void $(F)_01(size_t n, $T xi, $T *y, $T *z) {
     _$(F)_01<<<$BLK,$THR>>>(n,xi,y,z);
   }    
 }
